+++ Skin/Retail/AddOns/Blizzard_Channels.lua
@@ -6,9 +6,8 @@ if not private.isRetail then return end
 
 --[[ Core ]]
 local Aurora = private.Aurora
-local Base = Aurora.Base
 local Hook, Skin = Aurora.Hook, Aurora.Skin
-local Color, Util = Aurora.Color, Aurora.Util
+local Util = Aurora.Util
 
 do --[[ AddOns\Blizzard_Channels.lua ]]
     do --[[ ChannelButton.lua ]]
@@ -44,15 +43,12 @@ end
 
 do --[[ AddOns\Blizzard_Channels.xml ]]
     do --[[ ChannelButton.xml ]]
-        Skin.ChannelButtonBaseTemplate = private.nop
+        function Skin.ChannelButtonBaseTemplate(Button)
+            Skin.FrameTypeButton(Button)
+        end
         function Skin.ChannelButtonHeaderTemplate(Button)
             _G.hooksecurefunc(Button, "Update", Hook.ChannelButtonHeaderMixin.Update)
-
             Skin.ChannelButtonBaseTemplate(Button)
-            Button:SetNormalTexture("")
-            Button:SetHighlightTexture("")
-            Base.SetBackdrop(Button, Color.button)
-            Base.SetHighlight(Button)
 
             Button.Collapsed:SetAlpha(0)
             local minus = Button:CreateTexture(nil, "OVERLAY")
@@ -96,6 +92,7 @@ do --[[ AddOns\Blizzard_Channels.xml ]]
     end
     do --[[ ChannelList.xml ]]
         function Skin.ChannelListTemplate(ScrollFrame)
+            Util.Mixin(ScrollFrame.headerButtonPool, Hook.ObjectPoolMixin)
             Skin.UIPanelStretchableArtScrollBarTemplate(ScrollFrame.ScrollBar)
         end
     end
@@ -130,24 +127,22 @@ function private.AddOns.Blizzard_Channels()
     --       CreateChannelPopup       --
     ----====####$$$$%%%%$$$$####====----
     local CreateChannelPopup = _G.CreateChannelPopup
+    Skin.DialogBorderTemplate(CreateChannelPopup.BG)
+    local bg = CreateChannelPopup.BG:GetBackdropTexture("bg")
 
     CreateChannelPopup.Title:ClearAllPoints()
-    CreateChannelPopup.Title:SetPoint("TOPLEFT")
-    CreateChannelPopup.Title:SetPoint("BOTTOMRIGHT", CreateChannelPopup, "TOPRIGHT", 0, -private.FRAME_TITLE_HEIGHT)
+    CreateChannelPopup.Title:SetPoint("TOPLEFT", bg)
+    CreateChannelPopup.Title:SetPoint("BOTTOMRIGHT", bg, "TOPRIGHT", 0, -private.FRAME_TITLE_HEIGHT)
 
     CreateChannelPopup.Titlebar:Hide()
     CreateChannelPopup.Corner:Hide()
 
-    Skin.DialogBorderTemplate(CreateChannelPopup.BG)
     Skin.CreateChannelPopupEditBoxTemplate(CreateChannelPopup.Name)
     Skin.CreateChannelPopupEditBoxTemplate(CreateChannelPopup.Password)
     Skin.UICheckButtonTemplate(CreateChannelPopup.UseVoiceChat)
     Skin.UIPanelCloseButton(CreateChannelPopup.CloseButton)
     Skin.CreateChannelPopupButtonTemplate(CreateChannelPopup.OKButton)
-    CreateChannelPopup.OKButton:SetPoint("BOTTOMLEFT", 5, 5)
     Skin.CreateChannelPopupButtonTemplate(CreateChannelPopup.CancelButton)
-    CreateChannelPopup.CancelButton:ClearAllPoints()
-    CreateChannelPopup.CancelButton:SetPoint("BOTTOMRIGHT", -5, 5)
 
 
     ----====####$$$$%%%%%$$$$####====----